Renaming ATDCA to GOSP to better reflect the original paper
