+++ src/python_scripts/utils/parallel.py
@@ -37,7 +37,7 @@ from tqdm import tqdm
 from rasterio.windows import Window
 from dataclasses import dataclass
 
-from ..atdca.rastio import MultibandBlockReader
+from ..gosp.rastio import MultibandBlockReader
 from ..utils.math_utils import project_block_onto_subspace, block_l2_norms
